refactor: rename classes/Scripts -> classes/Maintenance namespace
Renames the Scripts namespace/module (CoverageChecker, CoverageComparator,
CoverageComparisonResult, CoverageResult, MaintenanceRepository,
PhpstanBaselineCounter, Contracts/MaintenanceRepositoryInterface) to
Maintenance to match its actual purpose (backlog/coverage maintenance
tooling), not general-purpose scripting. Sweeps all callers across
classes/, scripts/, root-level PHP, tests/, and phpunit.xml /
phpunit-mutation.xml testsuite registrations. Moves tests/Scripts ->
tests/Maintenance alongside.

Behavior-preserving; verified via full PHPUnit suite (6379 tests, 32200
assertions, 24 deprecations, 1 skipped — unchanged) and clean
composer run analyse / analyse:tests.
