+++ ibl5/classes/Maintenance/CoverageResult.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Scripts;
+namespace Maintenance;
 
 final readonly class CoverageResult
 {
